+++ tests/test_utils/test_imports.py
@@ -1,12 +1,10 @@
 from pathlib import Path
 
-from quickie import constants
-
 
 def test_import_from_path():
     from quickie.utils.imports import import_from_path
 
     root = Path.cwd()
-    path = root / constants.TASKS_PATH
+    path = root / "tests/__quickie_test"
     module = import_from_path(path)
-    assert module.__name__ == constants.TASKS_PATH.stem
+    assert module.__name__ == "__quickie_test"